Fixed a problem in creating markers during remeshing.
When an element has no markers in it, how to set the mattype of newly created
marker? This patch addresses this problem by looking at neighboring elements,
then counting the mattypes of their markers. The mattype of newly created is
determined by the probability of marker mattypes in neighboring elements.